Moved the hurricane tile creation process to the HHM Simple Engine
